Module vm_runtime::code_cache::module_adapter

Fetches code data from the blockchain.

Structs

FakeFetcher
A wrapper for a state with a list of pre-compiled modules.

ModuleFetcherImpl
A wrapper around State Store database for fetching code data stored on chain.

NullFetcher
A wrapper for an empty state with no code data stored.

Traits

ModuleFetcher
Trait that describes how the VM expects code data to be stored.