Trait vm_runtime::code_cache::module_cache::ModuleCache

pub trait ModuleCache<'alloc> {
    fn resolve_function_ref<'txn>(
        &'txn self, 
        caller_module: &LoadedModule, 
        idx: FunctionHandleIndex
    ) -> Result<Option<FunctionRef<'txn>>, VMInvariantViolation>;

    fn resolve_struct_def<'txn>(
        &'txn self, 
        module: &LoadedModule, 
        idx: StructDefinitionIndex, 
        gas_meter: &GasMeter
    ) -> VMResult<Option<StructDef>>;

    fn get_loaded_module(
        &self, 
        id: &CodeKey
    ) -> Result<Option<&'alloc LoadedModule>, VMInvariantViolation>;

    fn cache_module(
        &'alloc self, 
        module: CompiledModule
    ) -> Result<(), VMInvariantViolation>;

    fn reclaim_cached_module(
        &self, 
        v: Vec<LoadedModule>
    ) -> Result<(), VMInvariantViolation>;
}

Trait that describe a cache for modules. The idea is that this trait will in charge of loading resolving all dependencies of needed module from the storage.

Required methods

fn resolve_function_ref<'txn>(
&'txn self,
caller_module: &LoadedModule,
idx: FunctionHandleIndex
) -> Result<Option<FunctionRef<'txn>>, VMInvariantViolation>

Given a function handle index, resolve that handle into an internal representation of move function

fn resolve_struct_def<'txn>(
&'txn self,
module: &LoadedModule,
idx: StructDefinitionIndex,
gas_meter: &GasMeter
) -> VMResult<Option<StructDef>>

fn get_loaded_module(
&self,
id: &CodeKey
) -> Result<Option<&'alloc LoadedModule>, VMInvariantViolation>

Given a unique id for a module, return the resolved module structure if it is defined in cache.

fn cache_module(
&'alloc self,
module: CompiledModule
) -> Result<(), VMInvariantViolation>

fn reclaim_cached_module(
&self,
v: Vec<LoadedModule>
) -> Result<(), VMInvariantViolation>

TODO: Get rid of this function

Loading content...