Module vm_runtime::data_cache

Scratchpad for on chain values during the execution.

Structs

BlockDataCache
The wrapper around the StateVersionView for the block. It keeps track of the value that have been changed during execution of a block. It's effectively the write set for the block.

TransactionDataCache
Global cache for a transaction. Materializes Values from the RemoteCache and keeps an Rc to them. It also implements the opcodes that talk to storage and gives the proper guarantees of reference lifetime. Dirty objects are serialized and returned in make_write_set

Traits

RemoteCache
Trait for the StateVersionView or a mock implementation of the remote cache. Unit and integration tests should use this to mock implementations of "storage"