Struct vm_runtime::data_cache::TransactionDataCache

pub struct TransactionDataCache<'txn> { /* fields omitted */ }

Global cache for a transaction. Materializes Values from the RemoteCache and keeps an Rc to them. It also implements the opcodes that talk to storage and gives the proper guarantees of reference lifetime. Dirty objects are serialized and returned in make_write_set

Methods

impl<'txn> TransactionDataCache<'txn>

pub fn new(data_cache: &'txn dyn RemoteCache) -> Self

pub fn borrow_global(
&mut self,
ap: &AccessPath,
def: StructDef
) -> VMRuntimeResult<GlobalRef>

BorrowGlobal opcode cache implementation

pub fn resource_exists(
&mut self,
ap: &AccessPath,
def: StructDef
) -> (bool, u64)

Exists opcode cache implementation

pub fn move_resource_from(
&mut self,
ap: &AccessPath,
def: StructDef
) -> VMRuntimeResult<Local>

MoveFrom opcode cache implementation

pub fn move_resource_to(
&mut self,
ap: &AccessPath,
def: StructDef,
res: MutVal
) -> VMRuntimeResult<()>

MoveToSender opcode cache implementation

pub fn make_write_set(
&mut self,
to_be_published_modules: Vec<(CodeKey, Vec<u8>)>
) -> VMRuntimeResult<WriteSet>

Make a write set from the updated (dirty, deleted) global resources along with to-be-published modules. Consume the TransactionDataCache and must be called at the end of a transaction. This also ends up checking that reference count around global resources is correct at the end of the transactions (all ReleaseRef are properly called)

pub fn clear(&mut self)

Flush out the cache and restart from a clean state