Trait vm_runtime::loaded_data::function::FunctionReference

pub trait FunctionReference<'txn>: Sized + Clone {
    fn new(module: &'txn LoadedModule, idx: FunctionDefinitionIndex) -> Self;

    fn module(&self) -> &'txn LoadedModule;

    fn code_definition(&self) -> &'txn [Bytecode];

    fn local_count(&self) -> usize;

    fn arg_count(&self) -> usize;

    fn return_count(&self) -> usize;

    fn is_native(&self) -> bool;

    fn name(&self) -> &'txn IdentStr;

    fn signature(&self) -> &'txn FunctionSignature;
}

Trait that defines the internal representation of a move function.

Required methods

fn new(module: &'txn LoadedModule, idx: FunctionDefinitionIndex) -> Self

Create a new function reference to a module

fn module(&self) -> &'txn LoadedModule

Fetch the reference to the module where the function is defined

fn code_definition(&self) -> &'txn [Bytecode]

Fetch the code of the function definition

fn local_count(&self) -> usize

Return the number of locals for the function

fn arg_count(&self) -> usize

Return the number of input parameters for the function

fn return_count(&self) -> usize

Return the number of output parameters for the function

fn is_native(&self) -> bool

Return whether the function is native or not

fn name(&self) -> &'txn IdentStr

Return the name of the function

fn signature(&self) -> &'txn FunctionSignature

Returns the signature of the function

Loading content...