Struct vm_runtime::loaded_data::loaded_module::LoadedModule

pub struct LoadedModule {
    pub struct_defs_table: HashMap<Identifier, StructDefinitionIndex>,
    pub field_defs_table: HashMap<Identifier, FieldDefinitionIndex>,
    pub function_defs_table: HashMap<Identifier, FunctionDefinitionIndex>,
    pub function_defs: Vec<FunctionDef>,
    pub field_offsets: Vec<TableIndex>,
    // some fields omitted
}

Defines a loaded module in the memory. Currently we just store module itself with a bunch of reverse mapping that allows querying definition of struct/function by name.

Fields
struct_defs_table: HashMap<Identifier, StructDefinitionIndex>field_defs_table: HashMap<Identifier, FieldDefinitionIndex>function_defs_table: HashMap<Identifier, FunctionDefinitionIndex>function_defs: Vec<FunctionDef>field_offsets: Vec<TableIndex>
Methods

impl LoadedModule

pub fn new(module: VerifiedModule) -> Self

pub fn cached_struct_def_at(
&self,
idx: StructDefinitionIndex
) -> Option<StructDef>

Return a cached copy of the struct def at this index, if available.

pub fn cache_struct_def(&self, idx: StructDefinitionIndex, def: StructDef)

Cache this struct def at this location.

pub fn get_field_offset(
&self,
idx: FieldDefinitionIndex
) -> VMResult<TableIndex>

pub fn get_struct_def_index(
&self,
struct_name: &IdentStr
) -> VMResult<&StructDefinitionIndex>