Module vm_runtime::txn_executor

Processor for a single transaction.

Structs

ACCOUNT_MODULE
The ModuleId for the Account module

COIN_MODULE
The ModuleId for the LibraCoin module

EVENT_MODULE
The ModuleId for the Event

GAS_SCHEDULE_MODULE
The ModuleId for the gas schedule module

LIBRA_SYSTEM_MODULE
The ModuleId for the libra system module

TRANSACTION_FEE_DISTRIBUTION_MODULE
The ModuleId for the transaction fee distribution module

TransactionExecutor
A struct that executes one single transaction. 'alloc is the lifetime for the code cache, which is the argument type P here. Hence the P should live as long as alloc. 'txn is the lifetime of one single transaction.

VALIDATOR_CONFIG_MODULE
The ModuleId for the validator config

Functions

convert_txn_args
Convert the transaction arguments into move values.

execute_function_in_module
Execute the first function in a module