Trait vm_runtime::code_cache::module_cache::ModuleCache

pub trait ModuleCache<'alloc> {
    fn resolve_function_ref(
        &self, 
        caller_module: &LoadedModule, 
        idx: FunctionHandleIndex
    ) -> VMResult<Option<FunctionRef<'alloc>>>;

    fn resolve_struct_def(
        &self, 
        module: &LoadedModule, 
        idx: StructDefinitionIndex, 
        gas_meter: &GasMeter
    ) -> VMResult<Option<StructDef>>;

    fn get_loaded_module(
        &self, 
        id: &ModuleId
    ) -> VMResult<Option<&'alloc LoadedModule>>;

    fn cache_module(&self, module: VerifiedModule);

    fn reclaim_cached_module(&self, v: Vec<LoadedModule>);
}

Trait that describe a cache for modules. The idea is that this trait will in charge of loading resolving all dependencies of needed module from the storage.

Required methods

fn resolve_function_ref(
&self,
caller_module: &LoadedModule,
idx: FunctionHandleIndex
) -> VMResult<Option<FunctionRef<'alloc>>>

Given a function handle index, resolves that handle into an internal representation of move function.

Returns:

Ok(Some(FunctionRef)) if such function exists.

Ok(None) if such function doesn't exists.

Err(...) for a verification issue in a resolved dependency or VM invariant violation.

fn resolve_struct_def(
&self,
module: &LoadedModule,
idx: StructDefinitionIndex,
gas_meter: &GasMeter
) -> VMResult<Option<StructDef>>

Resolve a StructDefinitionIndex into a StructDef. This process will be recursive so we may charge gas on each recursive step.

Returns:

Ok(Some(StructDef)) if such struct exists.

Ok(None) if such function doesn't exists.

Err(...) for a verification or other issue in a resolved dependency, out of gas, or for a VM invariant violation.

fn get_loaded_module(
&self,
id: &ModuleId
) -> VMResult<Option<&'alloc LoadedModule>>

Resolve a ModuleId into a LoadedModule if the module has been cached already.

Returns:

Ok(Some(LoadedModule)) if such module exists.

Ok(None) if such module doesn't exists.

Err(...) for a verification issue in the module or for a VM invariant violation.

fn cache_module(&self, module: VerifiedModule)

fn reclaim_cached_module(&self, v: Vec<LoadedModule>)

Recache the list of previously resolved modules. Think of the cache as a generational cache and we need to move modules across generations.

Loading content...